@@ -6,6 +6,7 @@
     "test": "mocha lib/**/*.test.js"
   },
   "dependencies": {
+    "async": "2.3.0",
     "babel-core": "6.9.0",
     "babel-loader": "6.2.4",
     "babel-preset-es2015": "6.9.0",
@@ -17,6 +18,8 @@
     "file-loader": "0.8.5",
     "fs-extra": "2.1.2",
     "fuzzaldrin-plus": "^0.3.1",
+    "glob": "7.1.1",
+    "htmlparser2": "3.9.2",
     "imports-loader": "0.6.5",
     "invariant": "2.2.1",
     "jsdom": "9.11.0",
@@ -24,12 +27,14 @@
     "less-loader": "2.0.0",
     "lodash": "3.10.1",
     "marked": "0.3.3",
+    "megadoc-config-utils": "^6.0.0-beta.1",
+    "megadoc-corpus": "^6.0.0-beta.1",
     "multiline-slash": "2.0.0",
     "normalize.css": "^3.0.3",
     "prismjs": "1.5.0",
     "raw-loader": "0.5.1",
-    "react": "15.0.2",
-    "react-dom": "15.0.2",
+    "react": "15.4.2",
+    "react-dom": "15.4.2",
     "react-transclusion": "1.2.0",
     "style-loader": "0.13.1",
     "urijs": "1.18.0",
@@ -37,6 +42,7 @@
     "webpack": "1.12.15"
   },
   "devDependencies": {
+    "chai": "3.5.0",
     "connect": "3.4.1",
     "connect-modrewrite": "0.9.0",
     "json-loader": "0.5.1",
@@ -47,6 +53,7 @@
     "karma-sourcemap-loader": "0.3.3",
     "karma-webpack": "1.7.0",
     "mocha": "2.4.5",
+    "react-drill": "2.2.0",
     "react-hot-loader": "1.3.0",
     "serve-static": "^1.10.0",
     "webpack-dev-middleware": "1.6.1",
