@@ -189,7 +189,10 @@ function getByClassification(documentNode, klassifiers) {
 
 function getRemainingDocuments(documentNode, lists) {
   const usedUIDs = lists.reduce(function(map, list) {
-    list.forEach(x => map[x.uid] = true);
+    list.forEach(x => {
+      map[x.uid] = true;
+    });
+
     return map;
   }, {});
 
